Change the method of installing BrokenDownOptions
We need to change the implementation of a property in the
class `Options` (the class of the model's `_meta`).

We used to do it by changing the class from under the object once
normal model generation was over, but found out that that's too late.
So now we change the object's class on the way into the model class.